Refactor dbsp for enterprise backends.
- Moves certain storage functionality into a seperate crate
- Adjusts the way backend are found and created
