+++ crates/dbsp/src/circuit/tokio.rs
@@ -1,7 +1,4 @@
-use once_cell::sync::Lazy;
-use tokio::runtime::Runtime;
-
 /// The process running dbsp can share a single Tokio runtime.
 ///
 /// This is `pub` so it can be used in adapters too for IO.
-pub static TOKIO: Lazy<Runtime> = Lazy::new(|| Runtime::new().unwrap());
+pub use feldera_storage::tokio::TOKIO;
