+++ crates/dbsp/src/storage/buffer_cache/mod.rs
@@ -2,11 +2,10 @@
 
 /// A buffer-cache based on LRU eviction.
 mod cache;
-/// A file-backed buffer.
-mod fbuf;
+
+pub use feldera_storage::fbuf::{FBuf, FBufSerializer, LimitExceeded};
 
 pub use cache::{
     AtomicCacheCounts, AtomicCacheStats, BufferCache, CacheAccess, CacheCounts, CacheEntry,
     CacheStats,
 };
-pub use fbuf::{FBuf, FBufSerializer, LimitExceeded};
