+++ crates/feldera-types/src/config.rs
@@ -23,6 +23,7 @@ use serde::de::{self, MapAccess, Visitor};
 use serde::{Deserialize, Deserializer, Serialize};
 use serde_json::Value as JsonValue;
 use serde_yaml::Value as YamlValue;
+use std::fmt::Display;
 use std::path::Path;
 use std::{borrow::Cow, collections::BTreeMap};
 use utoipa::ToSchema;
@@ -174,6 +175,19 @@ pub enum StorageBackendConfig {
     /// `io_uring` is unavailable, as is often the case with cloud container
     /// systems.
     IoUring,
+
+    /// Object storage.
+    Object(ObjectStorageConfig),
+}
+
+impl Display for StorageBackendConfig {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            StorageBackendConfig::Default => write!(f, "default"),
+            StorageBackendConfig::IoUring => write!(f, "io_uring"),
+            StorageBackendConfig::Object(_) => write!(f, "object"),
+        }
+    }
 }
 
 /// Storage compression algorithm.
@@ -192,6 +206,74 @@ pub enum StorageCompression {
 
     /// Use [Snappy](https://en.wikipedia.org/wiki/Snappy_(compression)) compression.
     Snappy,
+}
+
+#[derive(Debug, Clone, Default, Eq, PartialEq, Serialize, Deserialize, ToSchema)]
+pub struct ObjectStorageConfig {
+    /// URL.
+    ///
+    /// The following URL schemes are supported:
+    ///
+    /// * S3:
+    ///   - `s3://<bucket>/<path>`
+    ///   - `s3a://<bucket>/<path>`
+    ///   - `https://s3.<region>.amazonaws.com/<bucket>`
+    ///   - `https://<bucket>.s3.<region>.amazonaws.com`
+    ///   - `https://ACCOUNT_ID.r2.cloudflarestorage.com/bucket`
+    /// * Google Cloud Storage:
+    ///   - `gs://<bucket>/<path>`
+    /// * Microsoft Azure Blob Storage:
+    ///   - `abfs[s]://<container>/<path>` (according to [fsspec](https://github.com/fsspec/adlfs))
+    ///   - `abfs[s]://<file_system>@<account_name>.dfs.core.windows.net/<path>`
+    ///   - `abfs[s]://<file_system>@<account_name>.dfs.fabric.microsoft.com/<path>`
+    ///   - `az://<container>/<path>` (according to [fsspec](https://github.com/fsspec/adlfs))
+    ///   - `adl://<container>/<path>` (according to [fsspec](https://github.com/fsspec/adlfs))
+    ///   - `azure://<container>/<path>` (custom)
+    ///   - `https://<account>.dfs.core.windows.net`
+    ///   - `https://<account>.blob.core.windows.net`
+    ///   - `https://<account>.blob.core.windows.net/<container>`
+    ///   - `https://<account>.dfs.fabric.microsoft.com`
+    ///   - `https://<account>.dfs.fabric.microsoft.com/<container>`
+    ///   - `https://<account>.blob.fabric.microsoft.com`
+    ///   - `https://<account>.blob.fabric.microsoft.com/<container>`
+    ///
+    /// Settings derived from the URL will override other settings.
+    pub url: String,
+
+    /// Additional options as key-value pairs.
+    ///
+    /// The following keys are supported:
+    ///
+    /// * S3:
+    ///   - `access_key_id`: AWS Access Key.
+    ///   - `secret_access_key`: AWS Secret Access Key.
+    ///   - `region`: Region.
+    ///   - `default_region`: Default region.
+    ///   - `endpoint`: Custom endpoint for communicating with S3,
+    ///     e.g. `https://localhost:4566` for testing against a localstack
+    ///     instance.
+    ///   - `token`: Token to use for requests (passed to underlying provider).
+    ///   - [Other keys](https://docs.rs/object_store/latest/object_store/aws/enum.AmazonS3ConfigKey.html#variants).
+    /// * Google Cloud Storage:
+    ///   - `service_account`: Path to the service account file.
+    ///   - `service_account_key`: The serialized service account key.
+    ///   - `google_application_credentials`: Application credentials path.
+    ///   - [Other keys](https://docs.rs/object_store/latest/object_store/gcp/enum.GoogleConfigKey.html).
+    /// * Microsoft Azure Blob Storage:
+    ///   - `access_key`: Azure Access Key.
+    ///   - `container_name`: Azure Container Name.
+    ///   - `account`: Azure Account.
+    ///   - `bearer_token_authorization`: Static bearer token for authorizing requests.
+    ///   - `client_id`: Client ID for use in client secret or Kubernetes federated credential flow.
+    ///   - `client_secret`: Client secret for use in client secret flow.
+    ///   - `tenant_id`: Tenant ID for use in client secret or Kubernetes federated credential flow.
+    ///   - `endpoint`: Override the endpoint for communicating with blob storage.
+    ///   - [Other keys](https://docs.rs/object_store/latest/object_store/azure/enum.AzureConfigKey.html#variants).
+    ///
+    /// Options set through the URL take precedence over those set with these
+    /// options.
+    #[serde(flatten)]
+    pub other_options: BTreeMap<String, String>,
 }
 
 /// Global pipeline configuration settings. This is the publicly
