+++ crates/storage/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "feldera-storage"
+version = "0.38.0"
+edition = "2021"
+
+[dependencies]
+feldera-types = { path = "../feldera-types", version = "0.38.0" }
+tokio = { version = "1.25.0", features = ["macros", "rt", "rt-multi-thread"] }
+libc = "0.2.153"
+rkyv = { version = "0.7.45", default-features = false, features = ["std", "size_64", "validation", "uuid"] }
+object_store = { version = "0.11.2", features = ["aws", "gcp", "azure", "http"] }
+serde = { version = "1.0", features = ["derive"] }
+thiserror = "2.0.11"
+uuid = { version = "1.12.1", features = ["v7"] }
+inventory = "0.3"
+tracing = "0.1.40"
+once_cell = "1.20.2"
+
+[target.'cfg(target_family = "unix")'.dependencies]
+nix = { version = "0.27.1", features = ["uio", "feature", "fs"] }
