+++ crates/storage/README.md
@@ -1,0 +1,4 @@
+## Feldera Storage
+
+Library that contains common types and traits for feldera storage backends used in dbsp.
+This crate is not intended to be used directly, but rather as a dependency for other crates.
