+++ crates/storage/src/fbuf.rs
